♻️ refactor(shared): extract AuthSessionStore from SettingsRepositoryImpl
